refactor mailer
