+++ .meteor/packages
@@ -4,16 +4,16 @@
 # 'meteor add' and 'meteor remove' will edit this file for you,
 # but you can also edit it by hand.
 
-meteor-base             # Packages every Meteor app needs to have
-mongo                   # The database Meteor supports right now
-reactive-var            # Reactive variable for tracker
-tracker                 # Meteor's client-side reactive programming library
+meteor-base@1.0.4             # Packages every Meteor app needs to have
+mongo@1.1.13                   # The database Meteor supports right now
+reactive-var@1.0.10            # Reactive variable for tracker
+tracker@1.1.0                 # Meteor's client-side reactive programming library
 
-standard-minifier-css   # CSS minifier run for production mode
-standard-minifier-js    # JS minifier run for production mode
-es5-shim                # ECMAScript 5 compatibility for older browsers.
-ecmascript              # Enable ECMAScript2015+ syntax in app code
-blaze-html-templates
+standard-minifier-css@1.2.0_1   # CSS minifier run for production mode
+standard-minifier-js@1.2.0_1    # JS minifier run for production mode
+es5-shim@4.6.14_1                # ECMAScript 5 compatibility for older browsers.
+ecmascript@0.5.8_1              # Enable ECMAScript2015+ syntax in app code
+blaze-html-templates@1.0.4
 
 mdg:validated-method
 aldeed:simple-schema
@@ -24,3 +24,4 @@ meteorhacks:ssr
 
 pathable-collections
 pathable-schema
+shell-server
