+++ .meteor/versions
@@ -1,100 +1,103 @@
-accounts-base@1.2.7
-accounts-password@1.1.8
-aldeed:collection2@2.9.1
-aldeed:collection2-core@1.1.1
-aldeed:schema-deny@1.0.1
-aldeed:schema-index@1.0.1
+accounts-base@1.2.13
+accounts-password@1.3.0
+aldeed:collection2@2.10.0
+aldeed:collection2-core@1.2.0
+aldeed:schema-deny@1.1.0
+aldeed:schema-index@1.1.0
 aldeed:simple-schema@1.5.3
-allow-deny@1.0.4
-autoupdate@1.2.7
+allow-deny@1.0.5
+autoupdate@1.3.11
 avital:mocha@2.1.0_10
-babel-compiler@6.9.1
-babel-runtime@0.1.11
-base64@1.0.8
-binary-heap@1.0.8
-blaze@2.1.7
-blaze-html-templates@1.0.3
-blaze-tools@1.0.8
-boilerplate-generator@1.0.8
-caching-compiler@1.0.3
-caching-html-compiler@1.0.5
-callback-hook@1.0.8
-check@1.2.1
-coffeescript@1.0.16
+babel-compiler@6.9.1_1
+babel-runtime@0.1.11_1
+base64@1.0.9
+binary-heap@1.0.9
+blaze@2.1.9
+blaze-html-templates@1.0.5
+blaze-tools@1.0.10
+boilerplate-generator@1.0.10
+caching-compiler@1.1.7_1
+caching-html-compiler@1.0.7
+callback-hook@1.0.9
+check@1.2.3
+coffeescript@1.2.4_2
 dburles:factory@1.1.0
 ddp@1.2.5
-ddp-client@1.2.5
-ddp-common@1.2.5
-ddp-rate-limiter@1.0.4
-ddp-server@1.2.6
+ddp-client@1.3.1_1
+ddp-common@1.2.6
+ddp-rate-limiter@1.0.5
+ddp-server@1.3.10_1
 deps@1.0.12
-diff-sequence@1.0.5
+diff-sequence@1.0.6
 dispatch:mocha-browser@0.0.3
 dispatch:mocha-core@0.0.2
-ecmascript@0.5.8
-ecmascript-runtime@0.3.14
-ejson@1.0.11
-email@1.0.12
-es5-shim@4.6.14
-geojson-utils@1.0.8
-hot-code-push@1.0.3
-html-tools@1.0.9
-htmljs@1.0.9
-http@1.1.4
-id-map@1.0.7
-jquery@1.11.8
-livedata@1.0.17
-localstorage@1.0.9
-logging@1.0.12
-mdg:validated-method@1.0.2
-mdg:validation-error@0.2.0
-meteor@1.2.17
-meteor-base@1.0.3
+ecmascript@0.5.8_1
+ecmascript-runtime@0.3.14_1
+ejson@1.0.12
+email@1.1.17_1
+es5-shim@4.6.14_1
+geojson-utils@1.0.9
+hot-code-push@1.0.4
+html-tools@1.0.11
+htmljs@1.0.11
+http@1.2.9_1
+id-map@1.0.8
+jquery@1.11.9
+livedata@1.0.18
+localstorage@1.0.11
+logging@1.1.15_1
+mdg:validated-method@1.1.0
+mdg:validation-error@0.5.1
+meteor@1.2.17_1
+meteor-base@1.0.4
 meteorhacks:cluster@1.6.9
 meteorhacks:ssr@2.2.0
-minifier-css@1.1.10
-minifier-js@1.1.11
-minimongo@1.0.14
-modules@0.7.6
-modules-runtime@0.7.6
-mongo@1.1.7
-mongo-id@1.0.4
-npm-bcrypt@0.8.7_1
-npm-mongo@1.4.43
-observe-sequence@1.0.11
-ordered-dict@1.0.7
+minifier-css@1.2.14_1
+minifier-js@1.2.14_1
+minimongo@1.0.17
+modules@0.7.6_1
+modules-runtime@0.7.6_1
+mongo@1.1.13
+mongo-id@1.0.5
+npm-bcrypt@0.9.1_1
+npm-mongo@2.2.11_1
+observe-sequence@1.0.13
+ordered-dict@1.0.8
 pathable-collections@0.0.1
 pathable-htmltostring@0.0.1
 pathable-schema@0.0.1
 pathable-utilities@0.0.1
-percolate:migrations@0.7.6
+percolate:migrations@0.9.8
 practicalmeteor:chai@2.1.0_1
 practicalmeteor:loglevel@1.2.0_2
 practicalmeteor:mocha-core@0.1.4
 practicalmeteor:sinon@1.14.1_2
-promise@0.8.4
+promise@0.8.7
 raix:eventemitter@0.1.3
-random@1.0.9
-rate-limit@1.0.4
-reactive-dict@1.1.6
-reactive-var@1.0.9
-reload@1.1.8
-retry@1.0.7
-routepolicy@1.0.10
-service-configuration@1.0.9
-session@1.1.4
-sha@1.0.7
-spacebars@1.0.11
-spacebars-compiler@1.0.11
-srp@1.0.8
-standard-minifier-css@1.0.5
-standard-minifier-js@1.0.5
-templating@1.1.8
-templating-tools@1.0.3
+random@1.0.10
+rate-limit@1.0.5
+reactive-dict@1.1.8
+reactive-var@1.0.10
+reload@1.1.10
+retry@1.0.8
+routepolicy@1.0.11
+service-configuration@1.0.10
+session@1.1.6
+sha@1.0.8
+shell-server@0.2.1
+spacebars@1.0.13
+spacebars-compiler@1.0.13
+srp@1.0.9
+standard-minifier-css@1.2.1
+standard-minifier-js@1.2.0_1
+templating@1.2.15
+templating-compiler@1.2.15
+templating-runtime@1.2.15
+templating-tools@1.0.5
 tmeasday:test-reporter-helpers@0.2.1
-tracker@1.0.13
-ui@1.0.11
+tracker@1.1.0
+ui@1.0.12
 underscore@1.0.9
-url@1.0.8
-webapp@1.2.8
+url@1.0.10
+webapp@1.3.11_1
 webapp-hashing@1.0.9
