+++ config/.env
@@ -1,4 +1,3 @@
-PACKAGES_DIR=/Users/peter/Development
-ROOT_URL=http://mailer.pathable.dev:3001
+ROOT_URL=http://mailer.pathable.dev:3002
 MONGO_URL=mongodb://localhost:27017/pathable
-PORT=3001
+PORT=3002
