--- imports/server/startup.js
@@ -1,1 +1,0 @@
-
