--- imports/server/templates/helpers.js
@@ -1,4 +1,0 @@
-// import { createLoginToken, createAuthorizedUrl } from 'meteor/pathable-utilities';
-
-
-// const authLink = (token) => ``;
