+++ imports/startup/server/register-api.js
@@ -1,0 +1,1 @@
+import '/imports/api/mail/methods.js';
