+++ package.json
@@ -13,6 +13,7 @@
     "pretest": "npm run lint --silent"
   },
   "dependencies": {
+    "bcrypt": "^0.8.7",
     "dotenv": "^2.0.0",
     "meteor-node-stubs": "~0.2.0",
     "mjml": "^2.3.3"
